These changes are reverting the changes integrated for revisions 510336 and 510281.  From the remarks on the dev mailing list, the basic idea is sound, we just need to re-visit the implementation a bit.
One "problem" that I couldn't resolve with this revert process was to restore the svn history for the DB2Dictionary.java part.  Somehow, the history for this part was lost when revision 510336 was integrated.  I have not found a means of restoring that history.  But, at least the contents are back to a known state.

git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/openjpa/trunk@512327 13f79535-47bb-0310-9956-ffa450edef68