OPENJPA-430 -- strip hungarian notation prefixes. Checking in patch for Ben Short, along with minor whitespace reformatting and a small tweak to reduce computation if branch logic in correctName() will not use the calculated name.
The initial patch included a test for fields that would end up having duplicate names after truncation. The code, however, does not do anything to resolve / avoid duplicates. To make it work, we would need to change MappingDefaultsImpl.correctName() to take a ValueMapping as an argument, and do a two-pass algorithm to check for other fields that would turn into duplicates. Even doing this will not be foolproof, as a duplicate column might come from a subclass or an embedded class. Since this option is not enabled by default anyways, and the duplicate error would become evident during testing / schema creation, I do not believe that it would be worthwhile to spend much time addressing this.

git-svn-id: https://svn.apache.org/repos/asf/openjpa/trunk@592319 13f79535-47bb-0310-9956-ffa450edef68