OPENJPA-2139 reset the BrokerFactory state if an Exception occured
This happens e.g. if the database is not available at the time the
first DB access is performed. This did lead to a complete hook up
until the whole application got restarted after the db reappeared.
This fix will now ensure that all the column-key caching and stuff
will be tried over again until the db is available again.


git-svn-id: https://svn.apache.org/repos/asf/openjpa/trunk@1311191 13f79535-47bb-0310-9956-ffa450edef68