OPENJPA-2861 fix unit test
Boolean is the correct return value.
This should actually have been returned for all dictionaries.
The problem is that we always only return the _internal_ representation.
The reason is because CASE/WHEN can be used to return values, but also as subquery.
Now imagine a database uses 0 and 1 for false and true. If CASE/WHEN is used
as subquery we really have to return 0/1 (number) because otherwise the WHERE clause
would not fit. When not executing the query on a Entity, we do not know the target type.
So there is probably no way we can later do a BooleanRepresentation call to switch to boolean.
And this would also break existing applications.